+++ CHANGELOG.md
@@ -2,24 +2,52 @@
 
 ## 0.3.0
 
+## Below is copied in from the old config repo, see: https://github.com/DEFRA/grant-config-woodland
+
+## 0.3.1
+
+### Patch Changes
+
+- ff5703f: Release the updates to the GAS config
+
+## 0.3.0
+
 ### Minor Changes
 
-- a5872ed: minor version bump to help align with existing config repo
+- 920a586: GSDS-390 move PA3 action from 'grant-config-woodland'
 
 ## 0.2.0
 
 ### Minor Changes
 
-- ce7edd1: remove temp change
+- 7b1be00: Add PA3 action config
 
 ## 0.1.1
 
 ### Patch Changes
 
-- 8e5307e: temp change to test grant config works
+- 7005d86: Update woodland config to latest. Run grants-ui journey tests when config changes
 
 ## 0.1.0
 
 ### Minor Changes
 
-- 6f9678a: Add grant configuration support to the repository
+- dee2ee2: Woodland config includes CW and GAS
+
+## 0.0.3
+
+### Patch Changes
+
+- 855ae92: Enable protected workflow
+
+## 0.0.2
+
+### Patch Changes
+
+- f8dc7d1: Add woodland grants-ui config
+
+## 0.0.1
+
+### Patch Changes
+
+- 54ca6e1: Create repo files